visual: extract define mode from vmenus.c to define.c
